AI-Alert-Assistance/backend/requirements.txt
flask==2.0.1
psycopg2-binary==2.9.1
kafka-python==2.0.2
python-json-logger==2.0.2
